Just playing around with print :)
This addresses #13, but I don't know if it adequately closes it.
This PR is the simplest thing to do, and was super fast to write.
Now, how can we improve it?  I'm not a big fan of how this always blocks.
If we're in Jupyter, can we print to the output of a cell asynchronously?