Fix race condition when updating output from other threads.
This applies when running in a notebook and may be a bug in ipywidgets.
`ipywidgets.Output.cleor_output()` does not wait until the output is cleared,
so we were printing new items, which got cleared when `clear_output()` finally
took effect.

Also, update calls to `ipywidgets.Output` to not use it as a context manager.
The new way is more thread-safe.